Moves validation module
This module was previous in the UI module, it now doesn't depend on DOM elements.
